refactor(calling): split messaging and voice-note services
