+++ src/yoyopod/communication/calling/__init__.py
@@ -1,7 +1,8 @@
 """Calling-domain implementations and app-facing helpers."""
 
 from yoyopod.communication.calling.history import CallHistoryEntry, CallHistoryStore
-from yoyopod.communication.calling.manager import VoIPManager, VoiceNoteDraft
+from yoyopod.communication.calling.manager import VoIPManager
+from yoyopod.communication.calling.voice_notes import VoiceNoteDraft
 
 __all__ = [
     "CallHistoryEntry",
